[PATCH] e1000: Remove duplex info from CTRL register dump
The duplex control register is used for setting the driver and is not
necessary for debug purposes.  The value of the duplex control register is
what the register's current value is and may not reflect the correct status
of te current connection.  That is what the duplex status register is used
for.  To keep from confusing the user, we are removing the duplex register
from the ethtool dump of the registers.

Signed-off-by: Jeff Kirsher <jeffrey.t.kirsher@intel.com>
Signed-off-by: Jesse Brandeburg <jesse.brandeburg@intel.com>
Signed-off-by: John Ronciak <john.ronciak@intel.com>
Signed-off-by: Jeff Garzik <jgarzik@pobox.com>