ethtool: Add option to flash firmware image from specified file, to a device.
This patch adds a new "-f" option to the ethtool utility to flash a
firmware image specified by a file, to a network device.  The filename
is passed to the network driver which will flash the image on the
chip using the request_firmware path.

The region "on the chip" to be flashed can be specified by an option.
It is up to the device driver to enumerate the region number passed
by ethtool, to the region to be flashed.

The default behavior is to flash all the regions on the chip.

Usage:
ethtool -f <interface name> <filename of firmware image>

ethtool -f <interface name> <filename of firmware image> [ REGION-NUMBER-TO-FLASH ]

Signed-off-by: Ajit Khaparde <ajitk@serverengines.com>
Signed-off-by: Jeff Garzik <jgarzik@redhat.com>