ethtool: Don't silently ignore speed/duplex when autoneg is on
When autonegotiation is enabled, drivers must determine link speed and
duplex through the autonegotiation process and will generally ignore
the speed and duplex specified in struct ethtool_cmd.  Currently, if
the user specifies autoneg on but does not specify the advertising
mask then:

- If the user specifies a recognised combination of speed and duplex,
  we set the advertising mask to the flag for that mode.  (Currently
  only one mode is recognised per combination of speed and duplex.)
- Otherwise, we advertise all recognised and supported modes.

But we should also set the advertising mask if autoneg is *already*
on.  Also, we should be able to limit the advertised modes separately
by speed and duplex.  For now, we just warn if we fail to do that.

Signed-off-by: Ben Hutchings <bhutchings@solarflare.com>