ethtool.8: Fix line-wrapping of options
Several options that take many keyword arguments are now formatted as
huge blocks of texts in the synopsis.  This is not very readable and
it results in a mixture of literal (must be typed) and non-literal
(must not be typed) hyphens.  Even with hyphenation, full justification
can insert extra space that look wrong in commands.

There are also minor problems with hyphenation in the detailed
descriptiona.

Therefore:
- Disable full justification and automatic hyphenation in the synopsis
- Use the .HP macro (paragraph with hanging indent) in the synopsis
  for all options that aren't already formatted with the .TP macro
- Write all literal hyphens as '\-' so they are not used for line-
  breaking

Signed-off-by: Ben Hutchings <bhutchings@solarflare.com>