ethtool: fix manpage so that it will display tables again
The current ethtool manpage is not displaying tables.  After trying to pass
the manpage through tbl I repeatedly saw the error:
tbl:ethtool.8.in:707: unrecognised format `x'
tbl:ethtool.8.in:707: giving up on this table

By dropping the 'x' the errors went away and when I built the manpage the
tables reappeared so I am assuming this is the correct approach.

[bwh: This format modifier improves table layout in groff 1.20.1, but
it's not worth losing backward compatibility.]

Signed-off-by: Alexander Duyck <alexander.h.duyck@intel.com>
[bwh: Removed two more 'x' format modifiers]
Signed-off-by: Ben Hutchings <bhutchings@solarflare.com>