Add RX packet classification interface
This patch was originally introduced as:
  [PATCH 1/3] [ethtool] Add rx pkt classification interface
  Signed-off-by: Santwona Behera <santwona.behera@sun.com>
  http://patchwork.ozlabs.org/patch/23223/

I removed the local caching of rules.

I dropped the use of regions as there were multiple issues found.

A network flow classifier is defined using the exact same syntax as
n-tuple, and the tool will correct for the fact that NFC uses the 1's
compliment of the n-tuple mask.

I also updated the ordering of new rules being added.  All new rules will
take the highest numbered open rule when no location is specified.

I split out the addition of bitops and the updates to documentation
into separate patches.  This makes the total patch size a bit more
manageable since the addition of NFC and the merging of it with
n-tuple were combined into this patch.

If we setup a rule and the device has the NTUPLE flag set we will
first try to use set_rx_ntuple.  If that fails with EOPNOTSUPP we then
will attempt to use the network flow classifier rule insertion.

Updated the output to make use of the updated network flow classifier
extensions that have been accepted into the kernel.

Merged the documentation update into this patch.  In addition the
documentation changes were made such that there is only one listing of the
individual options and they are all listed as optional.

Several fixes to address things such as the fact that we were
maintaining the table logic even though we only need it for displaying
all of the rules, or when adding a rule with no location specified.
As such all of the logic for deleting or finding rules in the table
has been removed.

Signed-off-by: Alexander Duyck <alexander.h.duyck@intel.com>
[bwh: Abbreviated the above commit message. Fixed a minor formatting
 error in the manual page additions.]
Signed-off-by: Ben Hutchings <bhutchings@solarflare.com>