Combine code to display supported and advertised link capabilities
Rename dump_advertised() to dump_link_caps(), and change
dump_supported() to call it instead of repeating nearly identical
code.  Add a separate prefix parameter and adjust indentation
so that the output is unchanged except for added lines.

Deleted unused parameter and variables.

As an immediate benefit, this adds the display of supported pause
frame use.

Signed-off-by: Ben Hutchings <bhutchings@solarflare.com>