Correct parameter types for ethtool -e and ethtool -E
All parameters to the underlying ethtool commands are unsigned, not
signed.  In particular, the 'magic' parameter to ethtool -E often has
the most significant bit set and users should not have to provide it
as a negative number.

For ethtool -E, the value to be written is 8-bit, not 32-bit.

Reported-by: Jens Rottmann <JRottmann@LiPPERTEmbedded.de>
Signed-off-by: Ben Hutchings <bhutchings@solarflare.com>