Rearrange definitions and remove unnecessary forward declarations
Most functions in ethtool.c are defined before use.  The major
exception is that the args array refers to a large number of functions
defined after it.  Move the args array and show_usage() functions down,
and move a few other functions up.

This leaves just one forward declaration in ethtool.c, which is
unavoidable because show_usage() and args refer to each other.

Signed-off-by: Ben Hutchings <bhutchings@solarflare.com>