Gracefully handle failure of register dump pretty-printer
If we have a pretty-printer matching the driver name for a register
dump, but it doesn't handle the reported dump version, then we
currently fail and print:

    Cannot dump registers: Success

Instead, fall back to hex dump.  Also, don't use perror() to explain a
failure of dump_regs().  errno is not meaningful in either failure
case.

Signed-off-by: Ben Hutchings <bhutchings@solarflare.com>