ixgbe: check mac type from ethtool_regs.version
This patch cleans up the mac type checks by using ethtool_regs.version
provided by the driver. This change eliminates the need to add device IDs
every time they are added to the driver.

Note that when using ethtool with this patch with a version of ixgbe that
does not provide the mac_type in ethtool_regs.version the register dump
may be incomplete. However this issue would've existed previously for
device IDs that were not added to ethtool.

Original patch and description by Emil Tantilov <emil.s.tantilov@intel.com>.

Signed-off-by: Ben Hutchings <bhutchings@solarflare.com>