Fix iteration count in sff8472_calibration()
We want to iterate over all elements of the various arrays of readings
(which all have the same dimensions - and maybe ought to be a single
array).  Unfortunately the iteration count is sizeof(sd->bias_cur) and
bias_cur has 2-byte elements, so we iterate over twice as many
elements as are really there.

Compiler-detected, and compile-tested only.

Signed-off-by: Ben Hutchings <ben@decadent.org.uk>
Signed-off-by: Ben Hutchings <bhutchings@solarflare.com>