test-cmdline: Disable test cases for --rxfh hash key parsing
test-cmdline does not mock the result of send_ioctl() but just treats
a call to send_ioctl() as indicating successful parsing.  Any parse
failure after this point cannot be tested for.  Currently, the hash
key string passed to --rxfh isn't parsed until after the device's key
size is known, so we can't test it tis way.

Signed-off-by: Ben Hutchings <ben@decadent.org.uk>